Sleep after issuing IPA server certificate
In some environments, the ipa role fails to run the second
ipa-server-install command to finish processing the externally issued
certificate.  It appears as if the time on the system and the start time
of the certificate are off and system time is before the start time of
the certificate.  Sleeping 10 seconds to give the system time to pass
the start time.

Also copying ipaserver-install.log after each step in case
troubleshooting is needed.