Keycloak updates for getting version and packages
Keycloak role checks github for latest version.  In some cases this
returns "null" and causes the role to fail.   Adding a loop to retry in
the event null is found.

The java version available has changed and needs to be updated from
java-17-openjdk-headless to java-21-openjdk-headless

Also make sure the service is started.  This is needed for VM setups in
other CIs.